refactor: 채팅카드모달 컴포넌트로 이름 변경
